requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.4
idna==2.10
multidict==4.7.6
pygooglechart==0.4.0
setuptools==41.2.0
websockets==8.1
yarl==1.4.2
youtube-dl==2020.6.16.1
APScheduler==3.6.3
mysql-connector==2.2.9
mysql-connector-python==8.0.21

